rename test to match pattern
